refactor(characters): extract Characters.Rolls
Fifth slice of #134. concept_roll! is the whole concept-roll concern; a
single-function module is deliberate — folding it into Effects or
Progression would blur their single-concern moduledocs for the sake of
avoiding a small file.
